More scaladoc; moved async.channel combinators in to pipe
