+++ core/jvm/src/main/scala/fs2/compress.scala
@@ -5,7 +5,9 @@ import java.util.zip.{DataFormatException, Deflater, Inflater}
 import scala.annotation.tailrec
 import scala.collection.mutable.ArrayBuffer
 
+/** Provides utilities for compressing/decompressing byte streams. */
 object compress {
+
   /**
     * Returns a `Pipe` that deflates (compresses) its input elements using
     * a `java.util.zip.Deflater` with the parameters `level`, `nowrap` and `strategy`.
